CSV file hospital quality
columns include the number of cases, figures for subcategories (e.g. rates of caesarean section) and mortality rates for specific clinical patterns and procedures. The indicators are intended to give the interested public an overview of activities conducted at hospitals largely financed by health insurance premiums and taxes, and to allow patients to make their choice of hospital based on facts. It is also intended to provide specialists at the hospitals concerned and outside with a basis for comparison between hospitals and inject a new aspect into discussions on quality.